third_party/blink/web_tests/virtual/async-load-data-url-fonts/fast/css/fontface-methods-expected.txt

CONSOLE WARNING: Failed to decode downloaded font: data:font/truetype;base64,
This is a testharness.js-based test.
PASS face1.load() should be fulfilled
PASS face2.load() should be rejected
PASS face3.load() should be fulfilled
PASS face4.load() should be rejected
PASS face5.load() should be fulfilled
Harness: the test ran to completion.
